configuration of CreditCard model class - tests, controller and services

+++ src/main/java/com/matcss/androidsdungeon/service/CreditCardService.java
@@ -2,7 +2,10 @@ package com.matcss.androidsdungeon.service;
 
 import com.matcss.androidsdungeon.interfaces.CRUDServices;
 import com.matcss.androidsdungeon.model.CreditCard;
+import com.matcss.androidsdungeon.model.Customer;
+import com.matcss.androidsdungeon.model.Product;
 import com.matcss.androidsdungeon.repository.CreditCardRepository;
+import com.matcss.androidsdungeon.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +16,9 @@ public class CreditCardService implements CRUDServices<CreditCard> {
 
     @Autowired
     private CreditCardRepository creditCardRepository;
+
+    @Autowired
+    private CustomerRepository customerRepository;
 
     @Override
     public List<CreditCard> findAll() {
@@ -29,21 +35,31 @@ public class CreditCardService implements CRUDServices<CreditCard> {
         return creditCardRepository.save(creditCard);
     }
 
+    public CreditCard create(int customerId, CreditCard obj) {
+        Customer customer = customerRepository.findCustomerByCustomerId(customerId);
+        if (customer == null)
+            return null;
+        obj.setCustomer(customer);
+
+        return creditCardRepository.save(obj);
+    }
+
     @Override
     public CreditCard update(int id, CreditCard creditCardBody) {
         CreditCard creditCard = findById(id);
 
         creditCard.setNumber(creditCardBody.getNumber());
-        creditCard.setCcv(creditCardBody.getCcv());
-        creditCard.setExpiration_date(creditCard.getExpiration_date());
+        creditCard.setCvv(creditCardBody.getCvv());
+        creditCard.setExpirationDate(creditCard.getExpirationDate());
         creditCard.setValidated(false);
 
         return creditCardRepository.save(creditCard);
     }
 
     @Override
-    public void delete(int id) {
+    public CreditCard delete(int id) {
         CreditCard creditCard = creditCardRepository.findCreditCardByCreditCardId(id);
-        creditCardRepository.delete(creditCard);
+        if(creditCard != null) creditCardRepository.delete(creditCard);
+        return creditCard;
     }
 }
